Get auth token from the GitHub CLI if possible (#31)
If no auth token is configured in the repository configuration, try
using the GitHub CLI command `gh auth token` to get it. If it returns a
token, use that instead of failing.

When running `jj spr init` on a repository, first check if the GitHub
CLI is already authenticated, and if so, offer to skip saving a new PAT.