Fix PR creation ordering and add :: operator support (#35)
* Fix `jj-spr diff -a` not creating PRs in correct order

When there is a stack of changes, `jj-spr diff -a` goes to create PRs
from newest to oldest, which is incorrect.

This fixes the ordering of PR when using a range. Additionally, it fixes
the revset operator used to be the ".." operator to match what `-a`
actually means. `jj-spr diff -a` means main..@-, which is all changes
excluding main and up to @-.

* Add support for inclusive operator

The old code when it was using :: operator to obtain the list of change,
seems to imply that the inclusive behavior is desirable. Therefore,
adding a way for range to be inclusive as well.